Query usage updates and more
This change started as an experiment to add a couple of extra Query patterns as examples. It revealed some weaknesses in how the layout was structured, which I was mostly aware of. But I knew I'd need to introduce a "no gap" block style until the Post Template block supports `blockGap`, so that was added.

Also, I dropped support for core patterns. They're messy anyway.