Zero out top margin on wrapper
Adds an explicit top margin on the wrapper block. This is necessary because the site editor sometimes puts its element styling at the top of the root container in `<style>` tags, breaking the usual `:first-child` rule of zeroing out the top margin of the first element. This often happens when navigating between templates in the editor.

This change also renames `page.html` to `singular.html`. This way, we have coverage for both pages and CPTs.